requirements.txt
apscheduler
apiclient==1.0.3
click==6.6
discovery==0.0.4
Flask==0.11.1
Flask-Assets==0.11
google-api-python-client==1.5.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2client==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
requests==2.5.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
urllib3==1.16
webassets==0.11.1
Werkzeug==0.11.10
wheel==0.24.0
